% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_leverage.R
\name{algo_leverage}
\alias{algo_leverage}
\title{algo_leverage function}
\usage{
algo_leverage(y, X, r = 0.1, sample_size = r, method = "leverage")
}
\arguments{
\item{y}{The response vector with n entries}

\item{X}{The covariates n by p matrix}

\item{r}{Percentage of subsamples to be drwan; default set to 0.1}

\item{sample_size}{Number of subsamples to consider for regression}

\item{method}{The type of subsampling: 'unif' for uniform-based, 'leverage' for leverage-score based}
}
\value{
The coefficient vector \code{beta}
}
\description{
The function implements algorithmic leveraging or linear regression using leverage score and uniform-based row subsampling.
It replicates the simulation setting of Fig 1 in Ma et Al's in https://onlinelibrary.wiley.com/doi/full/10.1002/wics.1324.
}
\examples{
algo_leverage(rnorm(500), matrix(rnorm(1000),nrow=500), method='unif')

}
