% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet_coord.R
\name{elnet_coord}
\alias{elnet_coord}
\title{Elastic Net Regression
Fits elastic net to data using coordinate descent algorithm using simulation setting described in HW1}
\usage{
elnet_coord()
}
\value{
The function returns the matrix of regression coefficients for each \eqn{\lambda} given as column vectors, the simulated design matrix \eqn{X}, and the simulated response matrix \eqn{Y}
}
\description{
Elastic Net Regression
Fits elastic net to data using coordinate descent algorithm using simulation setting described in HW1
}
