% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ols.R
\name{solve_ols}
\alias{solve_ols}
\title{solve_ols function}
\usage{
solve_ols(A, b, n_cores = 2, method, n_iter = 100)
}
\arguments{
\item{A}{the (square) matrix of coefficients of the linear system}

\item{b}{vector of values for each of the linear equation in the system}

\item{n_cores}{specifies the number of cores for parallel computing framework}

\item{method}{specifies the method to solve the linear system, "GS" for Gauss-Seidel, "Jacobi" for Jacobi}

\item{n_iter}{states the number of iterations to convergence/stop updates. Default is 100}
}
\value{
The solution vector \code{x} for linear system \code{Ax = b}
}
\description{
solve_ols function
}
\examples{
solve_ols(diag(rep(1, 3)), rep(1,3), n_cores = 2, method='GS', n_iter = 100)
}
